feat: 폐기물 배출 화면 수정 - 폐기물 배출 창고별로 리스트 출력되게 수정 - 애매한 함수들 전부 component로 이동 - util함수 전부 utils로 이동 - 리팩토링 몇몇 수행
