+++ app/src/main/java/com/example/myapplication/data/mock/MockBluetoothDevice.kt
@@ -1,0 +1,7 @@
+package com.example.myapplication.data.mock
+
+// 더미 블루투스 장치 모델
+data class MockBluetoothDevice(
+    val name: String,
+    val address: String
+)
