+++ app/src/main/java/com/example/myapplication/data/user/Beacon.kt
@@ -3,6 +3,6 @@ package com.example.myapplication.data.user
 data class Beacon(
     val id: Long,
     val macAddress: String,
-    val hospital: Hospital,
-    val isUsed: Boolean
+    val hospital: Hospital? = null,
+    val isUsed: Boolean = false
 )
