+++ app/src/main/java/com/example/myapplication/data/waste/SearchRequest.kt
@@ -3,6 +3,7 @@ package com.example.myapplication.data.waste
 data class SearchRequest(
     val wasteType: String? = null,
     val registrantName: String? = null,
+    val wasteStorageId: Long? = null,
     val selectedDate: String? = null,
     val selectedTime: String? = null,
     val combineDate: String? = null,
