+++ app/src/main/java/com/example/myapplication/ui/screen/RegisterScreen.kt
@@ -296,9 +296,3 @@ fun RegisterScreen(navController: NavController) {
         }
     }
 }
-
-@Preview(showBackground = true)
-@Composable
-fun RegisterScreenPreview() {
-    RegisterScreen(navController = rememberNavController())
-}
