+++ app/src/main/java/com/example/myapplication/ui/screen/WasteEditDialog.kt
@@ -2,13 +2,17 @@ package com.example.myapplication.ui.screen
 
 import android.util.Log
 import android.widget.Toast
+import androidx.compose.foundation.border
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.itemsIndexed
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowDropDown
 import androidx.compose.material3.AlertDialog
@@ -29,6 +33,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
@@ -98,18 +103,18 @@ fun WasteEditDialog(
             Column {
                 // 폐기물 종류 선택
                 Box(modifier = Modifier.fillMaxWidth()) {
-                    OutlinedTextField(
-                        value = wasteType,
-                        onValueChange = {},
-                        label = { Text("폐기물 종류") },
-                        readOnly = true,
-                        trailingIcon = {
-                            IconButton(onClick = { expandedType = true }) {
-                                Icon(Icons.Default.ArrowDropDown, contentDescription = "Dropdown")
-                            }
-                        },
-                        modifier = Modifier.fillMaxWidth()
-                    )
+                    Box(
+                        modifier = Modifier
+                            .fillMaxWidth()
+                            .clickable { expandedType = true }
+                            .border(1.dp, Color.Gray, RoundedCornerShape(4.dp))
+                            .padding(16.dp)
+                    ) {
+                        Text(
+                            text = wasteType,
+                            color = if (expandedType) Color.Gray else Color.Black
+                        )
+                    }
                     DropdownMenu(expanded = expandedType, onDismissRequest = { expandedType = false }) {
                         wasteTypes.forEach { type ->
                             DropdownMenuItem(
@@ -127,18 +132,18 @@ fun WasteEditDialog(
 
                 // 창고 선택
                 Box(modifier = Modifier.fillMaxWidth()) {
-                    OutlinedTextField(
-                        value = wasteStorage?.storageName.toString(),
-                        onValueChange = {},
-                        label = { Text("저장 창고") },
-                        readOnly = true,
-                        trailingIcon = {
-                            IconButton(onClick = { expandedStorage = true }) {
-                                Icon(Icons.Default.ArrowDropDown, contentDescription = "Dropdown")
-                            }
-                        },
-                        modifier = Modifier.fillMaxWidth()
-                    )
+                    Box(
+                        modifier = Modifier
+                            .fillMaxWidth()
+                            .clickable { expandedStorage = true }
+                            .border(1.dp, Color.Gray, RoundedCornerShape(4.dp))
+                            .padding(16.dp)
+                    ) {
+                        Text(
+                            text = wasteStorage?.storageName.toString(),
+                            color = if (expandedStorage) Color.Gray else Color.Black
+                        )
+                    }
                     DropdownMenu(expanded = expandedStorage, onDismissRequest = { expandedStorage = false }) {
                         wasteStorageList.forEach { storage ->
                             DropdownMenuItem(
