+++ app/src/main/java/com/example/myapplication/utils/DrawerContent.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication.ui.component
+package com.example.myapplication.utils
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
