+++ app/src/main/java/com/example/myapplication/utils/UtilFunctions.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication.ui.component
+package com.example.myapplication.utils
 
 import android.content.Context
 import android.os.Build
